Fixed VS2010 build errors.

+++ opensubdiv/osdutil/topology.cpp
@@ -266,7 +266,7 @@ PxOsdUtilSubdivTopology::ParseFromObjString(
         }
     }
 
-    numVertices = pointPositions->size()/3;
+    numVertices = (int)pointPositions->size()/3;
 
     return true;
 }
